feat(entities-plugins): expose freeform registry (#3415)
