+++ packages/entities/entities-plugins/src/components/free-form/shared/types.ts
@@ -1,4 +1,6 @@
-import type { UnionFieldSchema } from '../../../types/plugins/form-schema'
+import type { FormSchema, UnionFieldSchema } from '../../../types/plugins/form-schema'
+import type { FreeFormPluginData } from '../../../types/plugins/free-form'
+import type { PluginValidityChangeEvent } from '../../../types'
 import type { Component, ComponentPublicInstance, Ref, Slot } from 'vue'
 import { type LabelAttributes } from '@kong/kongponents'
 
@@ -201,6 +203,41 @@ export interface FieldRenderer {
   propsOverrides?: Record<string, unknown> | PropsOverridesFn
 }
 
+export type PluginConfigurationBaseProps<T extends Record<string, any> = Record<string, any>> = {
+  /** FreeForm Schema */
+  schema: FormSchema
+  /** The **initial** entire plugin model, never update */
+  model: T
+  /** Emits the final submission payload to the parent, the payload will be merged with the `formModel` but it has high override priority */
+  onFormChange: (value: Partial<T>, fields?: string[]) => void
+  /** FreeForm configuration */
+  formConfig?: FormConfig<T>
+  renderRules?: RenderRules
+  fieldRenderers?: FieldRenderer[]
+  pluginName: string
+  /** Konnect-managed Redis UI, from plugin form config */
+  isKonnectManagedRedisEnabled?: boolean
+}
+
+export type PluginFormLayoutProps<T extends FreeFormPluginData = FreeFormPluginData> = PluginConfigurationBaseProps<T> & {
+  onValidityChange?: (event: PluginValidityChangeEvent) => void
+  isEditing: boolean
+  /**
+   * Hide the built-in form/code switcher. Plugins that own a custom switcher
+   * (e.g. Datakit's flow/code control) should set this to true to avoid
+   * rendering duplicate controls into #plugin-form-page-actions.
+   */
+  hideEditorModeSwitcher?: boolean
+  /** Whether the plugin is being created for a portal developer */
+  developer?: boolean
+  generalInfoTitle?: string
+  generalInfoDescription?: string
+  pluginConfigTitle?: string
+  pluginConfigDescription?: string
+}
+
+export type PluginFormLayoutComponent<T extends FreeFormPluginData = FreeFormPluginData> = Component<PluginFormLayoutProps<T>>
+
 export interface PluginFormConfig {
   /**
    * Whether the plugin is experimental.
@@ -209,9 +246,10 @@ export interface PluginFormConfig {
   experimental?: boolean
   /**
    * Form-level custom component.
+    * Receives `PluginFormLayoutProps`.
    * @default CommonForm
    */
-  component: Component
+  component: PluginFormLayoutComponent<any>
   /**
    * Form-level rendering rules.
    */
