+++ packages/entities/entities-plugins/src/index.ts
@@ -6,6 +6,7 @@ import PluginCatalog from './components/PluginCatalog.vue'
 import PluginSelectGrid from './components/select/PluginSelectGrid.vue'
 import PluginSelectCard from './components/select/PluginSelectCard.vue'
 import PluginConfigCard from './components/PluginConfigCard.vue'
+import CommonForm from './components/free-form/Common/CommonForm.vue'
 import DynamicLayout from './components/free-form/shared/layout/DynamicLayout.vue'
 import PluginConfigurationForm from './components/free-form/shared/layout/PluginConfigurationForm.vue'
 import composables from './composables'
@@ -22,6 +23,7 @@ export {
   PluginSelectGrid,
   PluginSelectCard,
   PluginConfigCard,
+  CommonForm,
   DynamicLayout,
   PluginConfigurationForm,
   usePluginMetaData,
@@ -37,6 +39,14 @@ export type {
   PluginFormLayoutComponent,
   PluginFormLayoutProps,
 } from './components/free-form/shared/layout/provider'
+
+export {
+  pluginConfigRegistry,
+} from './components/free-form/shared/plugin-registry'
+
+export type {
+  ResolvedPluginFormConfig,
+} from './components/free-form/shared/plugin-registry'
 
 export * from './types'
